refact: move all app routes in routes dir
