+++ src/components/index.js
@@ -1,6 +1,4 @@
-export { ProtectedRoute } from './ProtectedRoute/ProtectedRoute';
 export { AdviseBox } from './AdviseBox/AdviseBox';
-export { AppRoutes } from './AppRoutes/AppRoutes';
 // Head
 export { UserInfo } from './UserInfo/UserInfo';
 export { Search } from './Search/Search';
