+++ src/routes/index.js
@@ -1,0 +1,3 @@
+export { AppRoutes } from './AppRoutes';
+export { MainPageRoutes } from './MainPageRoutes';
+export { ProtectedRoute } from './ProtectedRoute';
